c_importer_product: add tests common class
